Let the fun begin.
Rebranded man page (mostly).  There are still several errors and references to things like pesky alpha support.

Created a temporary hack to (hopefully) get BSD cpio working with the installation media.  If this works, we'll need to make some adjustments as to not waste memory and/or replace the mediaBufferSize function with something that makes sense.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@1113 818ba35e-4e59-e311-b59c-002590791c19