Latest developement version of libmport.
Changes:

It is now possible to create, install, and delete packages from a system
with the library.

The install primative now uses the progress callback system to give the user
feedback during the install.

Support for installing into a chroot envirement is included at the library
level.  This is useful if you have a GUI app and you don't want your entire
process stuck in the chroot.

Many function names have been changed in an effort to improve the library's
orginization.

Many bug fixes and subtle changes.

Todo:

Docs:
	library overview.
	function reference.
	mport bundle specification.

The merge primative.

The fetch primative.

The update primative.

Top level functions to do things like "fetch this bundle, its depends, and
then install them all."

Some of the sematics are still up in the air.  For example, gtk12 and
gtk2 both have a pkgname of gtk.  libmport doesn't allow two package with
the same name to be installed.

Hard links are not implemented in a mport bundle.

Much more testing and bug fixing is needed.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@1366 818ba35e-4e59-e311-b59c-002590791c19