PTY handling patch.
This is text from the FreeBSD advisory on the same issue:
If openpty(3) is called as non-root user the newly created
pseudo-terminal is world readable and writeable.  While this is
documented to be the case, script(1) still uses openpty(3) and
script(1) may be used by non-root users [CVE-2008-0217].

The ptsname(3) function incorrectly extracts two characters from the
name of a device node in /dev without verifying that it's actually
operating on a valid pty which the calling user owns.  pt_chown uses
the bad result from ptsname(3) to change ownership of a pty to the
user calling pt_chown [CVE-2008-0216].


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@1370 818ba35e-4e59-e311-b59c-002590791c19