$MidnightBSD$
Add some k8 errata hacks to powernow.

smist:
Use bus_dma to get a page in the first 4 GB.  Since the physical address
of the magic string is passed in a 32-bit register, we can't use high
memory in the PAE case.  This also eliminates a use of vtophys().

Obtained from: FreeBSD.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@1604 818ba35e-4e59-e311-b59c-002590791c19