ZFS is a file-system originally developed by Sun Microsystems.
The ZFS Intent Log ("ZIL") is a mechanism that gathers together in memory
transactions of writes, and is flushed onto disk when synchronous
semantics is necessary.  In the event of crash or power failure, the
log is examined and the uncommitted transaction would be replayed to
maintain the synchronous semantics.

When replaying setattr transaction, the replay code would set the
attributes with certain insecure defaults, when the logged
transaction did not touch these attributes.

A system crash or power fail would leave some file with mode set
to 07777.  This could leak sensitive information or cause privilege
escalation.

Based on FreeBSD security advisory SA-10:03.zfs (freebsd 7.x)


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@3221 818ba35e-4e59-e311-b59c-002590791c19