What a weird bug.  apparently some plists have @mode or @group commands with no arguments.  This is supposed to "reset" to the default state.  libmport can't handle this right now though.
As a workaround, i have added a case to check for data being null.   That means the argument is missing
@mode vs @mode 644 where data is 644 as stored in sqlite.

This allows us to uninstall phpMyAdmin.

The next question is what to do about this case. I think a default mode should be defined for installing files and directories for ports in libmport and we should die on weird edge cases for plists.  I don't think the package should have been created like this.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@3552 818ba35e-4e59-e311-b59c-002590791c19