Major refactor of mport tool code.
Use libmport to lookup entries in the index for fetching remote packages.  This means we can now create a magus run and load it on the ftp and have users remotely install mport packages!  (local install is a todo item)

Add a new update command which uninstalls a package and then fetches and installs a new one.  This is very "dumb" right now and it doesn't do updepends checks yet.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@3818 818ba35e-4e59-e311-b59c-002590791c19