Introduce support for preinit, init and fini arrays per the ELF standard.
This patch is based on work by John Marino for DragonFly.  This makes us the second BSD to support this feature.

From his comments:

For the main executable file, the .init_array and .fini_array sections
are handled by crt1, just like .init and .fini are.  In the case of
a statically linked binary, the .preinit_array section is also handled
by crt1.  The real-time linker handles the .init_array and .fini_array
sections for dynamically shared objects (libraries) and .preinit_array
for dynamically-linked binaries.  There are no .preinit_array sections
in the DSOs per standard.

These sections are described by the System V Application Binary Interface
http://www.sco.com/developers/gabi/latest/ch4.sheader.html#special_sections


In order for this patch to work, we also introduced several items from NetBSD needed for newer versions of GCC as well as some features used in this patch. (cdefs.h)

Finally, we're also through in manpages for cpu set which is totally unrelated, but an obmission from a previous patch for ULE.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@4637 818ba35e-4e59-e311-b59c-002590791c19