Introduce new functions:
mport_version() (public) to display bundle version and currently running os.

mport_get_osrelease() (private util) to get the os release from sysctl

Packate stubs now include the OS version the package was created for in the meta table as os_release key. This way, we know what OS version the package was built for. Fetching packages currently relies on a static define in the mport tool and this may go dynamic in the future.

It might be useful to install old packages in newer os versions in some scenarios such as jails.

Also, bump the fetch url to 0.5 for 0.5-CURRENT as we're starting to tinker with the package format.  The idea is to assume packages are for old releases if they don't include the meta[os_release] and allow a new mechanism to upgrade all packages for those built on the current release when we do major updates.


git-svn-id: http://svn.midnightbsd.org/svn/src/trunk@6024 818ba35e-4e59-e311-b59c-002590791c19